refactor(movex): :recycle: take the client dir out
