+++ libs/movex/src/lib/ConnectionToMasterResource.ts
