+++ libs/movex/src/lib/MovexBoundResource.ts
