+++ libs/movex/src/lib/MovexResource.ts
