+++ libs/movex/src/lib/MovexResourceObservable.spec.ts
