+++ libs/movex/src/lib/MovexResourceObservable.ts
