--- libs/movex/src/lib/client/index.ts
@@ -1,5 +1,0 @@
-export * from './Movex';
-export * from './init';
-export * from './MovexResource';
-export * from './MovexBoundResource';
-export * from './MovexFromDefintion';
