+++ libs/movex/src/lib/index.ts
@@ -1,2 +1,5 @@
-export * from './client/Movex';
-export * from './client/MovexFromDefintion';
+export * from './Movex';
+export * from './init';
+export * from './MovexResource';
+export * from './MovexBoundResource';
+export * from './MovexFromDefintion';
